js template code chage for WXPO on Toolkit
