+++ .vscode/tasks.json
@@ -31,7 +31,7 @@
     {
       "label": "Debug: Excel Desktop",
       "type": "npm",
-      "script": "start:desktop -- --app excel",
+      "script": "start -- --app excel",
       "presentation": {
         "clear": true,
         "panel": "dedicated",
@@ -41,7 +41,7 @@
     {
       "label": "Debug: Outlook Desktop",
       "type": "npm",
-      "script": "start:desktop -- --app outlook",
+      "script": "start -- --app outlook",
       "presentation": {
         "clear": true,
         "panel": "dedicated",
@@ -51,7 +51,7 @@
     {
       "label": "Debug: PowerPoint Desktop",
       "type": "npm",
-      "script": "start:desktop -- --app powerpoint",
+      "script": "start -- --app powerpoint",
       "presentation": {
         "clear": true,
         "panel": "dedicated",
@@ -61,7 +61,7 @@
     {
       "label": "Debug: Word Desktop",
       "type": "npm",
-      "script": "start:desktop -- --app word",
+      "script": "start -- --app word",
       "presentation": {
         "clear": true,
         "panel": "dedicated",
@@ -126,5 +126,22 @@
       },
       "problemMatcher": []
     },
+    {
+      "label": "Check OS",
+      "type": "shell",
+      "windows": {
+        "command": "echo 'Sideloading on Windows is supported'"
+      },
+      "linux": {
+        "command": "echo 'Sideloading on Linux is not supported' && exit 1"
+      },
+      "osx": {
+        "command": "echo 'Sideloading on Mac is not supported' && exit 1"
+      },
+      "presentation": {
+        "clear": true,
+        "panel": "dedicated",
+      },
+    }
   ]
 }
