+++ convertToSingleHost.js
@@ -2,12 +2,11 @@
 
 const convertTest = process.argv[3] === "convert-test";
 const fs = require("fs");
-const host = process.argv[2];
-const hosts = ["excel", "onenote", "outlook", "powerpoint", "project", "word"];
+const host = "outlook";
+const hosts = ["outlook", "word", "excel", "powerpoint"];
 const path = require("path");
 const util = require("util");
 const testPackages = [
-  "@babel/preset-typescript",
   "@types/mocha",
   "@types/node",
   "current-processes",
@@ -15,9 +14,7 @@ const testPackages = [
   "office-addin-mock",
   "office-addin-test-helpers",
   "office-addin-test-server",
-  "ts-loader",
   "ts-node",
-  "typescript",
 ];
 const readFileAsync = util.promisify(fs.readFile);
 const unlinkFileAsync = util.promisify(fs.unlink);
@@ -30,53 +27,14 @@ async function modifyProjectForSingleHost(host) {
   if (!hosts.includes(host)) {
     throw new Error(`'${host}' is not a supported host.`);
   }
-  await convertProjectToSingleHost(host);
+  await convertProjectToSingleHost();
   await updatePackageJsonForSingleHost(host);
   if (!convertTest) {
     await updateLaunchJsonFile();
   }
 }
 
-async function convertProjectToSingleHost(host) {
-  // copy host-specific manifest over manifest.xml
-  const manifestContent = await readFileAsync(`./manifest.${host}.xml`, "utf8");
-  await writeFileAsync(`./manifest.xml`, manifestContent);
-
-  // copy over host-specific taskpane code to taskpane.js
-  const srcContent = await readFileAsync(`./src/taskpane/${host}.js`, "utf8");
-  await writeFileAsync(`./src/taskpane/taskpane.js`, srcContent);
-
-  // delete all test files by default for now - eventually we want to convert the tests by default
-  if (convertTest && (host === "excel" || host === "word")) {
-    // copy over host-specific taskpane test code to test-taskpane.ts
-    const testTaskpaneContent = await readFileAsync(`./test/src/${host}-test-taskpane.ts`, "utf8");
-    const updatedTestTaskpaneContent = testTaskpaneContent.replace(
-      `../../src/taskpane/${host}`, 
-      `../../src/taskpane/taskpane.js`
-    );
-    await writeFileAsync(`./test/src/test-taskpane.ts`, updatedTestTaskpaneContent);
-
-    // update ui-test.ts to only run against specified host
-    const testContent = await readFileAsync(`./test/ui-test.ts`, "utf8");
-    const updatedTestContent = testContent.replace(`const hosts = ["Excel", "Word"]`, `const hosts = ["${host}"]`);
-    await writeFileAsync(`./test/ui-test.ts`, updatedTestContent);
-
-    // delete all host-specific test files after converting to single host
-    hosts.forEach(async function (host) {
-      if (host == "excel" || host == "word") {
-        await unlinkFileAsync(`./test/src/${host}-test-taskpane.ts`);
-      }
-    });
-  } else {
-    deleteFolder(path.resolve(`./test`));
-  }
-
-  // delete all host-specific files
-  hosts.forEach(async function (host) {
-    await unlinkFileAsync(`./manifest.${host}.xml`);
-    await unlinkFileAsync(`./src/taskpane/${host}.js`);
-  });
-
+async function convertProjectToSingleHost() {
   // delete the .github folder
   deleteFolder(path.resolve(`./.github`));
 
@@ -92,9 +50,6 @@ async function updatePackageJsonForSingleHost(host) {
   const packageJson = `./package.json`;
   const data = await readFileAsync(packageJson, "utf8");
   let content = JSON.parse(data);
-
-  // update 'config' section in package.json to use selected host
-  content.config["app_to_debug"] = host;
 
   // remove 'engines' section
   delete content.engines;
@@ -116,21 +71,19 @@ async function updatePackageJsonForSingleHost(host) {
     }
   });
 
-  if (!convertTest) {
-    // remove test-related scripts
-    Object.keys(content.scripts).forEach(function (key) {
-      if (key.includes("test")) {
-        delete content.scripts[key];
-      }
-    });
+  // remove test-related scripts
+  Object.keys(content.scripts).forEach(function (key) {
+    if (key.includes("test")) {
+      delete content.scripts[key];
+    }
+  });
 
-    // remove test-related packages
-    Object.keys(content.devDependencies).forEach(function (key) {
-      if (testPackages.includes(key)) {
-        delete content.devDependencies[key];
-      }
-    });
-  }
+  // remove test-related packages
+  Object.keys(content.devDependencies).forEach(function (key) {
+    if (testPackages.includes(key)) {
+      delete content.devDependencies[key];
+    }
+  });
 
   // write updated json to file
   await writeFileAsync(packageJson, JSON.stringify(content, null, 2));
@@ -172,7 +125,6 @@ async function deleteSupportFiles() {
   await unlinkFileAsync("./convertToSingleHost.js");
   await unlinkFileAsync(".npmrc");
   await unlinkFileAsync("package-lock.json");
-  await unlinkFileAsync("tsconfig.json"); // used only for test file building in js project
 }
 
 /**
