+++ package.json
@@ -7,7 +7,6 @@
   },
   "license": "MIT",
   "config": {
-    "app_to_debug": "excel",
     "app_type_to_debug": "desktop",
     "dev_server_port": 3000
   },
@@ -23,15 +22,12 @@
     "dev-server": "webpack serve --mode development",
     "lint": "office-addin-lint check",
     "lint:fix": "office-addin-lint fix",
+    "signin": "office-addin-dev-settings m365-account login",
+    "signout": "office-addin-dev-settings m365-account logout",
     "prettier": "office-addin-lint prettier",
-    "start": "office-addin-debugging start manifest.xml",
-    "start:desktop": "office-addin-debugging start manifest.xml desktop",
-    "start:desktop:outlook": "office-addin-debugging start manifest.outlook.xml desktop --app outlook",
-    "start:web": "office-addin-debugging start manifest.xml web",
+    "start": "office-addin-debugging start manifest.json",
     "stop": "office-addin-debugging stop manifest.xml",
     "test": "npm run test:unit && npm run test:e2e",
-    "test:e2e": "mocha -r ts-node/register test/end-to-end/*.ts",
-    "test:unit": "mocha -r ts-node/register test/unit/*.test.ts",
     "validate": "office-addin-manifest validate manifest.xml",
     "watch": "webpack --mode development --watch"
   },
