+++ src/taskpane/taskpane.js
@@ -1,6 +1,86 @@
-import "./excel";
-import "./onenote";
-import "./outlook";
-import "./powerpoint";
-import "./project";
-import "./word";
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+ * See LICENSE in the project root for license information.
+ */
+
+/* global document, Office */
+
+Office.onReady((info) => {
+    if (info.host === Office.HostType.Outlook) {
+      document.getElementById("sideload-msg").style.display = "none";
+      document.getElementById("app-body").style.display = "flex";
+      document.getElementById("run").onclick = runOutlook;
+    }
+    if(info.host === Office.HostType.Word){
+      document.getElementById("sideload-msg").style.display = "none";
+      document.getElementById("app-body").style.display = "flex";
+      document.getElementById("run").onclick = runWord;
+    }
+    if (info.host === Office.HostType.Excel) {
+      document.getElementById("sideload-msg").style.display = "none";
+      document.getElementById("app-body").style.display = "flex";
+      document.getElementById("run").onclick = runExcel;
+    }
+    if (info.host === Office.HostType.PowerPoint) {
+      document.getElementById("sideload-msg").style.display = "none";
+      document.getElementById("app-body").style.display = "flex";
+      document.getElementById("run").onclick = runPowerPoint;
+    }
+  });
+
+export async function runOutlook() {
+    /**
+     * Insert your Outlook code here
+     */
+
+    const item = Office.context.mailbox.item;
+    document.getElementById("item-subject").innerHTML = "<b>Subject:</b> <br/>" + item.subject;
+}  
+
+export async function runWord() {
+return Word.run(async (context) => {
+    /**
+     * Insert your Word code here
+     */
+
+    // insert a paragraph at the end of the document.
+    const paragraph = context.document.body.insertParagraph("Hello World", Word.InsertLocation.end);
+
+    // change the paragraph color to blue.
+    paragraph.font.color = "blue";
+
+    await context.sync();
+});
+}
+
+export async function runExcel() {
+    try {
+      await Excel.run(async (context) => {
+        /**
+         * Insert your Excel code here
+         */
+        const range = context.workbook.getSelectedRange();
+  
+        // Read the range address
+        range.load("address");
+  
+        // Update the fill color
+        range.format.fill.color = "yellow";
+  
+        await context.sync();
+        console.log(`The range address was ${range.address}.`);
+      });
+    } catch (error) {
+      console.error(error);
+    }
+  }
+
+export async function runPowerPoint() {
+/**
+ * Insert your PowerPoint code here
+ */
+const options = { coercionType: Office.CoercionType.Text };
+
+await Office.context.document.setSelectedDataAsync(" ", options);
+await Office.context.document.setSelectedDataAsync("Hello World!", options);
+}
